Add READ-CHAR Based Tests
This gives a facility for reading one character at a time from either
a smart terminal or a redirected input device.

It is truly a *character* based interface, so it is interpreting bytes
as UTF-8.  A separate READ-BYTE interface will be needed for reading
one byte at a time from files.

The test that's added reverses the characters of the input stream