Move TRY to the KEEP in %tac.r
Changes to TRY mean it no longer returns a BLANK! as an input which
can silently be tolerated by functions as wanting them to be a no-op.
The TRY is now moved to suppress a specialized error from that
function that happens when it returns null.

In the case of KEEP, it does not yet return an error on NULL, but
a null isotope.  This is likely going to change.  Either way, moving
the TRY to the keep (as TRY KEEP vs. KEEP TRY) fixes a problem at
this moment.